requirements.txt
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
pandas-ta>=0.3.14b
niftystocks>=0.0.2
requests>=2.31
beautifulsoup4>=4.12
transformers>=4.36
torch>=2.1
scipy>=1.11
streamlit>=1.30
plotly>=5.18
tabulate>=0.9
anthropic>=0.40
openai>=1.10
python-dotenv>=1.0
